[PLXUTILS-151] FileUtils.deleteDirectory() attempts to traverse symlinks, then fails when it encounters a file it can't delete

+++ src/main/java/org/codehaus/plexus/util/FileUtils.java
@@ -1562,6 +1562,14 @@ public class FileUtils
             return;
         }
 
+        /* try delete the directory before its contents, which will take
+         * care of any directories that are really symbolic links.
+         */
+        if ( directory.delete() )
+        {
+        	return;
+        }
+
         cleanDirectory( directory );
         if ( !directory.delete() )
         {
